pyproject.toml
[project]
name = "dbhub-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "langchain>=1.2.17",
    "langchain-google-genai>=4.2.2",
    "langchain-openai>=1.2.1",
    "langgraph>=1.1.10",
    "mcp>=1.27.0",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.46.0",
]
